Hector Lugo Summer UCR REU Portion
These files encompass data pertaining to an alteration of PageRank, clustering, multi-view graphing, and various other graphical representations of the dataset. My primary focus is on depicting accidents that occur within the same county and are associated with the same company, showcasing various implementations of this aspect. The data also includes details on the number of people injured or killed, analyzed using different algorithms. Additionally, it integrates weather data to highlight potential correlations with the frequency of accidents.